cloud_storage/tests: separate out remote_partition fuzz tests
...and run them as an _rpfixture instead of an _rpunit.

This should make them more reliable, by running them sequentially
instead of in parallel with lots of other unit tests.  It also helps
to speed up unit test runs.
